Include Images in Conversion Format
Some Facebook posts contain an image representation if their contents
has an image, video, or link. For this kind of post, the image
representation is now included in the conversion.

The background color of the conversion is reverted to its original
blue color. Instead, when the conversion is copied to the clipboard,
the background color will temporarily be changed to white and changed
back once the copy has been made.

In addition to inclusion of images, the conversion now includes a link
to view the original post on Facebook, and a signature indicating that
the conversion was made by this Google Chrome extension.